Restructuring parser code
* Move syntax structs into a separate package
* Move all module parsing into a single impl file
* More docs
