+++ src/format/text/module/mod.rs
@@ -1,0 +1,3 @@
+pub mod syntax;
+pub mod macros;
+pub mod parse;
