+++ src/spec/format/mod.rs
@@ -1,0 +1,7 @@
+use crate::format::text::Parser;
+
+impl <'a> Parser<'a> {
+
+    fn parse_spec_test(&mut self) -> Result<SpecTest> {
+    }
+}
